third commit

+++ src/routes/files.js
@@ -11,13 +11,21 @@ dotenv.config();
 
 export default (app) => {
   app.get('/file/download/:id', async (request, response) => {
+    if (!request.user) {
+      response.status(403).end();
+      return;
+    }
     const { id } = request.params;
-    const file = await app.dataSource.getRepository('File').findBy({ id });
-    const { path } = file[0];
+    const file = await app.dataSource.getRepository('File').findOneBy({ id });
+    const { path } = file;
     response.sendFile(path);
   });
 
   app.get('/file/list', async (request, response) => {
+    if (!request.user) {
+      response.status(403).end();
+      return;
+    }
     const listSize = request.query.list_size || 10;
     const page = request.query.page || 1;
     const offset = listSize * (page - 1);
@@ -26,13 +34,25 @@ export default (app) => {
   });
 
   app.get('/file/:id', async (request, response) => {
+    if (!request.user) {
+      response.status(403).end();
+      return;
+    }
     const { id } = request.params;
-    const file = await app.dataSource.getRepository('File').findBy({ id });
+    const file = await app.dataSource.getRepository('File').findOneBy({ id });
     response.send(file);
   });
 
   const uploadDirPath = join(__dirname, '..', '..', process.env.UPLOAD_DIR_NAME);
   app.post('/file/upload', multer({ dest: uploadDirPath }).single('file'), async (request, response) => {
+    if (!request.user) {
+      response.status(403).end();
+      return;
+    }
+    if (!request.file) {
+      response.end();
+      return;
+    }
     const {
       originalname,
       mimetype,
@@ -51,16 +71,32 @@ export default (app) => {
   });
 
   app.delete('/file/delete/:id', async (request, response) => {
+    if (!request.user) {
+      response.status(403).end();
+      return;
+    }
     const { id } = request.params;
-    const file = await app.dataSource.getRepository('File').findBy({ id });
-    const { path } = file[0];
+    const file = await app.dataSource.getRepository('File').findOneBy({ id });
+    if (!file) {
+      response.end();
+      return;
+    }
+    const { path } = file;
     unlink(path, () => {});
     await app.dataSource.getRepository('File').delete({ id });
     response.end();
   });
 
   app.put('/file/update/:id', multer({ dest: uploadDirPath }).single('file'), async (request, response) => {
+    if (!request.user) {
+      response.status(403).end();
+      return;
+    }
     const { id } = request.params;
+    if (!request.file) {
+      response.end();
+      return;
+    }
     const {
       originalname,
       mimetype,
